move JSProperty from pyscript.web.dom to pyscript.web.elements
